Revert "rework api"
