--- .env
@@ -1,6 +1,0 @@
-DB_HOST=localhost
-DB_PORT=5432
-DB_USER=postgres
-DB_PASSWORD=12345
-DB_NAME=cashlog_db
-PORT=3000
