+++ .gitignore
@@ -1,3 +1,3 @@
 .idea/
 .vscode/
-node_modules/
+frontend/node_modules/
