+++ backend/internal/storage/DataBase.go
@@ -1,0 +1,20 @@
+package DataBase
+
+import (
+	"log"
+
+	"gorm.io/driver/postgres"
+	"gorm.io/gorm"
+)
+
+var DB *gorm.DB
+
+func PostgreSQL() {
+	dsn := "host=localhost user=postgres password=12345 dbname=postgres port=5432 sslmode=disable"
+	var err error
+	DB, err = gorm.Open(postgres.Open(dsn), &gorm.Config{})
+	if err != nil {
+		log.Fatal("Failed to connect to database:", err)
+	}
+	log.Println("Database connected successfully")
+}
