--- config/cfg.go
@@ -1,26 +1,0 @@
-package config
-
-import (
-	"fmt"
-	"log"
-	"os"
-
-	"github.com/joho/godotenv"
-)
-
-func LoadConfig() {
-	err := godotenv.Load(".env")
-	if err != nil {
-		log.Fatal("Error loading .env")
-	}
-}
-
-func PsqlCfg() string {
-	dbUser := os.Getenv("DB_USER")
-	dbPass := os.Getenv("DB_PASSWORD")
-	dbHost := os.Getenv("DB_HOST")
-	dbPort := os.Getenv("DB_PORT")
-	dbName := os.Getenv("DB_NAME")
-
-	return fmt.Sprintf("postgres://%s:%s@%s:%s/%s?sslmode=disable", dbUser, dbPass, dbHost, dbPort, dbName)
-}
