--- public/styles.css
@@ -1,272 +1,0 @@
-:root {
-	/* Light theme */
-	--bg-color: #f5f5f7;
-	--text-color: #1d1d1f;
-	--primary-color: #0071e3;
-	--secondary-color: #86868b;
-	--table-bg: white;
-	--table-header-bg: #f5f5f7;
-	--border-color: #e5e5e7;
-	--shadow-color: rgba(0, 0, 0, 0.03);
-	--error-color: #ff3b30;
-	--success-color: #34c759;
-}
-
-.dark-theme {
-	--bg-color: #1d1d1f;
-	--text-color: #f5f5f7;
-	--primary-color: #3a3a3c;
-	--secondary-color: #636366;
-	--table-bg: #2c2c2e;
-	--table-header-bg: #3a3a3c;
-	--border-color: #3a3a3c;
-	--shadow-color: rgba(0, 0, 0, 0.2);
-	--error-color: #ff453a;
-	--success-color: #30d158;
-}
-
-* {
-	margin: 0;
-	padding: 0;
-	box-sizing: border-box;
-	font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen,
-		Ubuntu, Cantarell, sans-serif;
-	transition: background-color 0.3s ease, color 0.3s ease,
-		border-color 0.3s ease;
-}
-
-body {
-	background-color: var(--bg-color);
-	color: var(--text-color);
-	padding: 2rem;
-	min-height: 100vh;
-}
-
-.container {
-	max-width: 1200px;
-	margin: 0 auto;
-}
-
-/* Header styles */
-.header {
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	margin-bottom: 2rem;
-}
-
-h1,
-h2,
-h3 {
-	color: var(--text-color);
-	font-weight: 600;
-}
-
-h1 {
-	font-size: 24px;
-}
-
-/* Button styles */
-.btn {
-	padding: 0.8rem 1.5rem;
-	border: none;
-	border-radius: 12px;
-	background-color: var(--primary-color);
-	color: white;
-	font-weight: 500;
-	cursor: pointer;
-	transition: opacity 0.2s, background-color 0.3s;
-}
-
-.btn:hover {
-	opacity: 0.8;
-}
-
-.btn-secondary {
-	background-color: var(--secondary-color);
-}
-
-.btn-danger {
-	background-color: var(--error-color);
-}
-
-/* Form styles */
-.form-group {
-	margin-bottom: 1.5rem;
-}
-
-label {
-	display: block;
-	margin-bottom: 0.5rem;
-	color: var(--text-color);
-	font-weight: 500;
-}
-
-input[type='text'],
-input[type='email'],
-input[type='password'],
-textarea {
-	width: 100%;
-	padding: 0.8rem;
-	border: 1px solid var(--border-color);
-	border-radius: 12px;
-	background-color: var(--table-bg);
-	color: var(--text-color);
-	font-size: 1rem;
-	transition: border-color 0.3s;
-}
-
-textarea {
-	resize: vertical;
-	min-height: 100px;
-}
-
-input:focus,
-textarea:focus {
-	outline: none;
-	border-color: var(--primary-color);
-	box-shadow: 0 0 0 2px rgba(0, 113, 227, 0.2);
-}
-
-.error-message {
-	color: var(--error-color);
-	margin-top: 1rem;
-}
-
-/* Notes list styles */
-.notes-list {
-	display: grid;
-	grid-template-columns: repeat(auto-fill, minmax(280px, 1fr));
-	gap: 1.5rem;
-	margin-top: 2rem;
-}
-
-.note-card {
-	background-color: var(--table-bg);
-	border-radius: 18px;
-	padding: 1.5rem;
-	box-shadow: 0 2px 8px var(--shadow-color);
-	transition: transform 0.2s ease;
-}
-
-.note-card:hover {
-	transform: translateY(-2px);
-}
-
-.note-title {
-	font-size: 1.2rem;
-	margin-bottom: 0.5rem;
-	color: var(--text-color);
-}
-
-.note-content {
-	color: var(--text-color);
-	opacity: 0.8;
-	margin-bottom: 1rem;
-}
-
-.note-actions {
-	display: flex;
-	gap: 0.8rem;
-}
-
-/* Modal styles */
-.modal {
-	display: none;
-	position: fixed;
-	top: 0;
-	left: 0;
-	width: 100%;
-	height: 100%;
-	background-color: rgba(0, 0, 0, 0.5);
-	z-index: 1000;
-}
-
-.modal-content {
-	background-color: var(--table-bg);
-	margin: 5% auto;
-	padding: 2rem;
-	border-radius: 18px;
-	width: 90%;
-	max-width: 500px;
-	box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
-}
-
-.modal-header {
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	margin-bottom: 1.5rem;
-}
-
-.close {
-	color: var(--text-color);
-	font-size: 2rem;
-	cursor: pointer;
-	transition: opacity 0.2s;
-}
-
-.close:hover {
-	opacity: 0.7;
-}
-
-.modal-actions {
-	display: flex;
-	gap: 1rem;
-	justify-content: flex-end;
-	margin-top: 2rem;
-}
-
-/* Auth pages styles */
-.auth-container {
-	max-width: 400px;
-	margin: 0 auto;
-	padding: 2rem;
-	background-color: var(--table-bg);
-	border-radius: 18px;
-	box-shadow: 0 2px 8px var(--shadow-color);
-}
-
-.auth-header {
-	text-align: center;
-	margin-bottom: 2rem;
-}
-
-.auth-footer {
-	text-align: center;
-	margin-top: 1.5rem;
-}
-
-/* Theme toggle */
-.theme-toggle {
-	position: fixed;
-	bottom: 2rem;
-	right: 2rem;
-	width: 50px;
-	height: 50px;
-	border-radius: 50%;
-	background-color: var(--primary-color);
-	color: white;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	cursor: pointer;
-	box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-	border: none;
-}
-
-/* Responsive adjustments */
-@media (max-width: 768px) {
-	body {
-		padding: 1rem;
-	}
-
-	.notes-list {
-		grid-template-columns: 1fr;
-	}
-
-	.modal-content {
-		width: 95%;
-		padding: 1.5rem;
-	}
-}
